refactor: Create Products component to manage rendering of Product components
