+++ 12.webDeveloper/project/src/components/Home.jsx
@@ -1,24 +1,8 @@
-import { useBdContext } from "../context/bd"
-import { Product } from "./Product"
-import { allImages } from "../assets"
+import { Products } from "./Products"
 
 
 export const Home = () => {
-    const handleBd = useBdContext()
-    console.log(handleBd);
-
-    const allProducts = Object.keys(handleBd.mockDb)
-
-
     return (<>
-        {allProducts.map(element => {
-            return <Product
-                currency={handleBd.mockDb[element].currency}
-                price={handleBd.mockDb[element].price}
-                title={element}
-                description={handleBd.mockDb[element].description}
-                image={allImages[element]}
-                measure={handleBd.mockDb[element].measure} />
-        })}
+        <Products />
     </>)
 }
